Refactor swig modules to package
