+++ shyft/orchestration2/base_config.py
@@ -59,7 +59,7 @@ class BaseConfig(object):
 
     @property
     def model_api(self):
-        from shyft import pt_gs_k
+        from shyft.api import pt_gs_k
         return getattr(pt_gs_k, self.model_config.parameters['model']['model_api'])
 
     def __init__(self, config_file, config_section, **overrides):
